Add files via upload
added main.py and added functionality to build an inventory (as an array of items) from the db.  I also added a subcategories attribute to the item class in item_cart_class.py as a dictionary, which should allow us to add as many subcategories as we want to any item we want without having to have new classes for them.  I also fixed a mistake with Spiderman vs SpiderMan in the db